requirements.txt
redis
ttrpg_core
slack_pybot
py_kanka_wrapper
pytest
slackclient
